+++ app/src/main/java/woowacourse/shopping/shopping/recentview/RecentViewedRecyclerAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.shopping
+package woowacourse.shopping.shopping.recentview
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
